+++ sources/xcode/Jasonelle/JLKernel/JLKernel/JLKernel.h
@@ -22,7 +22,7 @@
 //  with this file, You can obtain one at
 //
 //  <https://mozilla.org/MPL/2.0/>.
-//
+
 
 #import <Foundation/Foundation.h>